Delete useless code
